Ajout de controle des données saisies

+++ api/contacts/deleteContact.php
@@ -4,9 +4,7 @@
 header("Content-Type: application/json; charset=UTF-8");
 
 // On inclut nos class pour pouvoir y accéder
-require_once('./../../config/Database.php');
-require_once('./../../managers/ContactManager.php');
-require_once('./../../models/Contact.php');
+require_once('./../../config/init.php');
 
 // On instancie la base de données et on s'y connecte
 $database = new Database();
@@ -21,19 +19,23 @@ $contact = new Contact();
 // On récupère les données envoyées
 $datas = json_decode(file_get_contents("php://input"));
 
-// On affecte les attributs de notre contact
-$contact->hydrate($datas);
-
-// On modifie le contact
-if ($contactManager->delete($contact)) {
-    $result['message'] = "Le contact a bien été supprimé";
-    echo json_encode($result);
+// On vérifie les données reçues
+if (!empty($datas->id) && is_numeric($datas->id)) {
+    // On affecte les attributs de notre contact
+    $contact->hydrate($datas);
+    
+    // On modifie le contact
+    if ($contactManager->delete($contact)) {
+        $result['message'] = "Le contact a bien été supprimé";
+        echo json_encode($result);
+    } else {
+        $result['message'] = "Le contact n'a pas pu être supprimé";
+        echo json_encode($result);
+    }
 } else {
-    $result['message'] = "Le contact n'a pas pu être supprimé";
+    $result['message'] = "Les informations données ne sont incorrects";
     echo json_encode($result);
 }
 
-// Afficher un message de succès
-// ...
 
 ?>
